+++ src/types.ts
@@ -1,4 +1,4 @@
-export type SnapshotType = "auto" | "manual";
+export type SnapshotType = "auto" | "manual" | "synced";
 
 export type Snapshot = {
   id: string;
@@ -6,6 +6,7 @@ export type Snapshot = {
   name?: string;
   type: SnapshotType;
   items: string[];
+  playbackPosition?: number; // milliseconds
 };
 
 export type AutoMode = "timer" | "on-change";
@@ -22,9 +23,10 @@ export type Settings = {
   promptManualBeforeReplace: boolean;
   language?: string;
   settingsCollapsed: boolean;
+  syncedSnapshotId?: string;
 };
 
-export type BadgeVariant = "default" | "accent" | "version" | "channel";
+export type BadgeVariant = "default" | "accent" | "version" | "channel" | "active-sync";
 
 export type ButtonTone = "primary" | "danger" | "subtle" | "default";
 
@@ -41,7 +43,5 @@ export type QueueUpdateEventData = {
   nextUp: Spicetify.PlayerTrack[];
 };
 
-export type QueueUpdateEvent = Event & {
-  type: "queue_update";
-  data?: QueueUpdateEventData;
-};
+export type QueueUpdateEvent = Event & { type: "queue_update"; data: QueueUpdateEventData };
+export type OnProgressEvent = Event & { type: "onprogress"; data: number };